+++ lib/src/in_app_webview/in_app_webview_settings.dart
@@ -1174,7 +1174,7 @@ class InAppWebViewSettings_ {
   ///
   ///**NOTE for iOS**: available on iOS 15.4+.
   ///
-  ///**NOTE for MacOS**: available on MacOS 12.0+.
+  ///**NOTE for MacOS**: available on MacOS 12.3+.
   ///
   ///**Supported Platforms/Implementations**:
   ///- iOS
@@ -1199,7 +1199,7 @@ class InAppWebViewSettings_ {
   ///
   ///**NOTE for iOS**: available on iOS 15.4+.
   ///
-  ///**NOTE for MacOS**: available on MacOS 12.0+.
+  ///**NOTE for MacOS**: available on MacOS 12.3+.
   ///
   ///**Supported Platforms/Implementations**:
   ///- iOS